fragment auto blocking added
