+++ blocks/footer/footer.js
@@ -1,6 +1,5 @@
 import { getMetadata } from '../../scripts/aem.js';
-import { loadFragment } from '../fragment/fragment.js';
-
+import { loadFragment } from '../../scripts/scripts.js';
 /**
  * loads and decorates the footer
  * @param {Element} block The footer block element
